requirements.txt
discord.py

python >= 3.8